# zstd: extend executeSimple with history support