@WojciechMula 2022-03-24
Impact of commit https://github.com/klauspost/compress/pull/542/commits/3a351249bf595b30a6dd32f5bb774b0611c5a5a4 (comparison with the previous commit from this branch).

Can't tell it's significantly better or worse, except few cases.

```
benchmark                                                                 old ns/op     new ns/op     delta
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-16                            3275140       3197659       -2.37%
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-16                        688157        688801        +0.09%
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-16                         10174938      9977297       -1.94%
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-16                           7464937       7488839       +0.32%
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-16                         2315824       2322484       +0.29%
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-16                          2971254       2984982       +0.46%
BenchmarkDecoder_DecoderSmall/html_x_4.zst-16                             1218236       1199092       -1.57%
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-16                       204441        203376        -0.52%
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-16                       126309        126411        +0.08%
BenchmarkDecoder_DecoderSmall/urls.10K.zst-16                             9711496       9721413       +0.10%
BenchmarkDecoder_DecoderSmall/html.zst-16                                 729087        731648        +0.35%
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-16                        69411         69318         -0.13%
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-16                               401512        404052        +0.63%
BenchmarkDecoder_DecodeAll/geo.protodata.zst-16                           83962         83687         -0.33%
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-16                            1155677       1168934       +1.15%
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-16                              860139        873312        +1.53%
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-16                            299327        301737        +0.81%
BenchmarkDecoder_DecodeAll/alice29.txt.zst-16                             375086        380545        +1.46%
BenchmarkDecoder_DecodeAll/html_x_4.zst-16                                228614        228758        +0.06%
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-16                          21867         21759         -0.49%
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-16                          11299         11300         +0.01%
BenchmarkDecoder_DecodeAll/urls.10K.zst-16                                1107666       1099319       -0.75%
BenchmarkDecoder_DecodeAll/html.zst-16                                    93401         93670         +0.29%
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-16                           8788          8786          -0.02%
BenchmarkDecoder_DecodeAllFiles/Mark.Twain-Tom.Sawyer.txt/fastest-16      933956        940400        +0.69%
BenchmarkDecoder_DecodeAllFiles/Mark.Twain-Tom.Sawyer.txt/default-16      904977        908869        +0.43%
BenchmarkDecoder_DecodeAllFiles/Mark.Twain-Tom.Sawyer.txt/better-16       857884        864611        +0.78%
BenchmarkDecoder_DecodeAllFiles/Mark.Twain-Tom.Sawyer.txt/best-16         902840        959689        +6.30%
BenchmarkDecoder_DecodeAllFiles/e.txt/fastest-16                          9180          9194          +0.15%
BenchmarkDecoder_DecodeAllFiles/e.txt/default-16                          218821        217876        -0.43%
BenchmarkDecoder_DecodeAllFiles/e.txt/better-16                           186824        186077        -0.40%
BenchmarkDecoder_DecodeAllFiles/e.txt/best-16                             152161        150956        -0.79%
BenchmarkDecoder_DecodeAllFiles/fse-artifact3.bin/fastest-16              3326          3268          -1.74%
BenchmarkDecoder_DecodeAllFiles/fse-artifact3.bin/default-16              3008          2954          -1.80%
BenchmarkDecoder_DecodeAllFiles/fse-artifact3.bin/better-16               3917          3958          +1.05%
BenchmarkDecoder_DecodeAllFiles/fse-artifact3.bin/best-16                 10700         10746         +0.43%
BenchmarkDecoder_DecodeAllFiles/gettysburg.txt/fastest-16                 4753          4665          -1.85%
BenchmarkDecoder_DecodeAllFiles/gettysburg.txt/default-16                 7352          7277          -1.02%
BenchmarkDecoder_DecodeAllFiles/gettysburg.txt/better-16                  7367          7333          -0.46%
BenchmarkDecoder_DecodeAllFiles/gettysburg.txt/best-16                    7250          7770          +7.17%
BenchmarkDecoder_DecodeAllFiles/html.txt/fastest-16                       59693         60146         +0.76%
BenchmarkDecoder_DecodeAllFiles/html.txt/default-16                       61358         61251         -0.17%
BenchmarkDecoder_DecodeAllFiles/html.txt/better-16                        57739         57833         +0.16%
BenchmarkDecoder_DecodeAllFiles/html.txt/best-16                          62846         61222         -2.58%
BenchmarkDecoder_DecodeAllFiles/pi.txt/fastest-16                         9191          9209          +0.20%
BenchmarkDecoder_DecodeAllFiles/pi.txt/default-16                         220163        219331        -0.38%
BenchmarkDecoder_DecodeAllFiles/pi.txt/better-16                          183516        182773        -0.40%
BenchmarkDecoder_DecodeAllFiles/pi.txt/best-16                            151235        152682        +0.96%
BenchmarkDecoder_DecodeAllFiles/pngdata.bin/fastest-16                    27742         27312         -1.55%
BenchmarkDecoder_DecodeAllFiles/pngdata.bin/default-16                    31985         31819         -0.52%
BenchmarkDecoder_DecodeAllFiles/pngdata.bin/better-16                     25521         25129         -1.54%
BenchmarkDecoder_DecodeAllFiles/pngdata.bin/best-16                       33093         32938         -0.47%
BenchmarkDecoder_DecodeAllFiles/sharnd.out/fastest-16                     9187          9186          -0.01%
BenchmarkDecoder_DecodeAllFiles/sharnd.out/default-16                     9181          9184          +0.03%
BenchmarkDecoder_DecodeAllFiles/sharnd.out/better-16                      9176          9178          +0.02%
BenchmarkDecoder_DecodeAllFiles/sharnd.out/best-16                        9202          9210          +0.09%
BenchmarkDecoder_DecodeAllFilesP/Mark.Twain-Tom.Sawyer.txt/fastest-16     125946        124304        -1.30%
BenchmarkDecoder_DecodeAllFilesP/Mark.Twain-Tom.Sawyer.txt/default-16     129010        128391        -0.48%
BenchmarkDecoder_DecodeAllFilesP/Mark.Twain-Tom.Sawyer.txt/better-16      118192        117074        -0.95%
BenchmarkDecoder_DecodeAllFilesP/Mark.Twain-Tom.Sawyer.txt/best-16        123044        123888        +0.69%
BenchmarkDecoder_DecodeAllFilesP/e.txt/fastest-16                         1033          1036          +0.29%
BenchmarkDecoder_DecodeAllFilesP/e.txt/default-16                         35850         35520         -0.92%
BenchmarkDecoder_DecodeAllFilesP/e.txt/better-16                          28095         27881         -0.76%
BenchmarkDecoder_DecodeAllFilesP/e.txt/best-16                            18661         18632         -0.16%
BenchmarkDecoder_DecodeAllFilesP/fse-artifact3.bin/fastest-16             614           717           +16.80%
BenchmarkDecoder_DecodeAllFilesP/fse-artifact3.bin/default-16             725           655           -9.70%
BenchmarkDecoder_DecodeAllFilesP/fse-artifact3.bin/better-16              687           639           -7.00%
BenchmarkDecoder_DecodeAllFilesP/fse-artifact3.bin/best-16                983           985           +0.15%
BenchmarkDecoder_DecodeAllFilesP/gettysburg.txt/fastest-16                925           932           +0.70%
BenchmarkDecoder_DecodeAllFilesP/gettysburg.txt/default-16                1362          1372          +0.73%
BenchmarkDecoder_DecodeAllFilesP/gettysburg.txt/better-16                 1351          1346          -0.37%
BenchmarkDecoder_DecodeAllFilesP/gettysburg.txt/best-16                   1591          1586          -0.31%
BenchmarkDecoder_DecodeAllFilesP/html.txt/fastest-16                      13069         13083         +0.11%
BenchmarkDecoder_DecodeAllFilesP/html.txt/default-16                      13489         13615         +0.93%
BenchmarkDecoder_DecodeAllFilesP/html.txt/better-16                       12777         12862         +0.67%
BenchmarkDecoder_DecodeAllFilesP/html.txt/best-16                         13337         13475         +1.03%
BenchmarkDecoder_DecodeAllFilesP/pi.txt/fastest-16                        1036          1032          -0.39%
BenchmarkDecoder_DecodeAllFilesP/pi.txt/default-16                        35567         35691         +0.35%
BenchmarkDecoder_DecodeAllFilesP/pi.txt/better-16                         27402         27815         +1.51%
BenchmarkDecoder_DecodeAllFilesP/pi.txt/best-16                           18755         18689         -0.35%
BenchmarkDecoder_DecodeAllFilesP/pngdata.bin/fastest-16                   4964          4908          -1.13%
BenchmarkDecoder_DecodeAllFilesP/pngdata.bin/default-16                   5051          5038          -0.26%
BenchmarkDecoder_DecodeAllFilesP/pngdata.bin/better-16                    4404          4347          -1.29%
BenchmarkDecoder_DecodeAllFilesP/pngdata.bin/best-16                      4081          4033          -1.18%
BenchmarkDecoder_DecodeAllFilesP/sharnd.out/fastest-16                    1040          1031          -0.87%
BenchmarkDecoder_DecodeAllFilesP/sharnd.out/default-16                    1034          1038          +0.39%
BenchmarkDecoder_DecodeAllFilesP/sharnd.out/better-16                     1034          1038          +0.39%
BenchmarkDecoder_DecodeAllFilesP/sharnd.out/best-16                       1041          1049          +0.77%
BenchmarkDecoder_DecodeAllParallel/kppkn.gtb.zst-16                       67962         68051         +0.13%
BenchmarkDecoder_DecodeAllParallel/geo.protodata.zst-16                   19570         19321         -1.27%
BenchmarkDecoder_DecodeAllParallel/plrabn12.txt.zst-16                    199859        197860        -1.00%
BenchmarkDecoder_DecodeAllParallel/lcet10.txt.zst-16                      151725        153752        +1.34%
BenchmarkDecoder_DecodeAllParallel/asyoulik.txt.zst-16                    48884         48180         -1.44%
BenchmarkDecoder_DecodeAllParallel/alice29.txt.zst-16                     67054         67496         +0.66%
BenchmarkDecoder_DecodeAllParallel/html_x_4.zst-16                        59590         58551         -1.74%
BenchmarkDecoder_DecodeAllParallel/paper-100k.pdf.zst-16                  8079          8055          -0.30%
BenchmarkDecoder_DecodeAllParallel/fireworks.jpeg.zst-16                  1266          1262          -0.32%
BenchmarkDecoder_DecodeAllParallel/urls.10K.zst-16                        179567        179464        -0.06%
BenchmarkDecoder_DecodeAllParallel/html.zst-16                            19411         19355         -0.29%
BenchmarkDecoder_DecodeAllParallel/comp-data.bin.zst-16                   1996          1985          -0.55%

benchmark                                                                 old MB/s     new MB/s     speedup
BenchmarkDecoder_DecoderSmall/kppkn.gtb.zst-16                            450.23       461.14       1.02x
BenchmarkDecoder_DecoderSmall/geo.protodata.zst-16                        1378.62      1377.33      1.00x
BenchmarkDecoder_DecoderSmall/plrabn12.txt.zst-16                         378.86       386.37       1.02x
BenchmarkDecoder_DecoderSmall/lcet10.txt.zst-16                           457.34       455.88       1.00x
BenchmarkDecoder_DecoderSmall/asyoulik.txt.zst-16                         432.43       431.19       1.00x
BenchmarkDecoder_DecoderSmall/alice29.txt.zst-16                          409.49       407.61       1.00x
BenchmarkDecoder_DecoderSmall/html_x_4.zst-16                             2689.79      2732.74      1.02x
BenchmarkDecoder_DecoderSmall/paper-100k.pdf.zst-16                       4007.02      4028.00      1.01x
BenchmarkDecoder_DecoderSmall/fireworks.jpeg.zst-16                       7796.32      7789.99      1.00x
BenchmarkDecoder_DecoderSmall/urls.10K.zst-16                             578.36       577.77       1.00x
BenchmarkDecoder_DecoderSmall/html.zst-16                                 1123.60      1119.66      1.00x
BenchmarkDecoder_DecoderSmall/comp-data.bin.zst-16                        469.78       470.41       1.00x
BenchmarkDecoder_DecodeAll/kppkn.gtb.zst-16                               459.06       456.18       0.99x
BenchmarkDecoder_DecodeAll/geo.protodata.zst-16                           1412.40      1417.05      1.00x
BenchmarkDecoder_DecodeAll/plrabn12.txt.zst-16                            416.95       412.22       0.99x
BenchmarkDecoder_DecodeAll/lcet10.txt.zst-16                              496.15       488.66       0.98x
BenchmarkDecoder_DecodeAll/asyoulik.txt.zst-16                            418.20       414.86       0.99x
BenchmarkDecoder_DecodeAll/alice29.txt.zst-16                             405.48       399.66       0.99x
BenchmarkDecoder_DecodeAll/html_x_4.zst-16                                1791.67      1790.54      1.00x
BenchmarkDecoder_DecodeAll/paper-100k.pdf.zst-16                          4682.77      4706.12      1.00x
BenchmarkDecoder_DecodeAll/fireworks.jpeg.zst-16                          10894.56     10892.92     1.00x
BenchmarkDecoder_DecodeAll/urls.10K.zst-16                                633.84       638.66       1.01x
BenchmarkDecoder_DecodeAll/html.zst-16                                    1096.34      1093.20      1.00x
BenchmarkDecoder_DecodeAll/comp-data.bin.zst-16                           463.80       463.91       1.00x
BenchmarkDecoder_DecodeAllFiles/Mark.Twain-Tom.Sawyer.txt/fastest-16      415.40       412.55       0.99x
BenchmarkDecoder_DecodeAllFiles/Mark.Twain-Tom.Sawyer.txt/default-16      428.70       426.86       1.00x
BenchmarkDecoder_DecodeAllFiles/Mark.Twain-Tom.Sawyer.txt/better-16       452.23       448.72       0.99x
BenchmarkDecoder_DecodeAllFiles/Mark.Twain-Tom.Sawyer.txt/best-16         429.71       404.26       0.94x
BenchmarkDecoder_DecodeAllFiles/e.txt/fastest-16                          10893.36     10876.76     1.00x
BenchmarkDecoder_DecodeAllFiles/e.txt/default-16                          457.01       458.99       1.00x
BenchmarkDecoder_DecodeAllFiles/e.txt/better-16                           535.28       537.43       1.00x
BenchmarkDecoder_DecodeAllFiles/e.txt/best-16                             657.22       662.46       1.01x
BenchmarkDecoder_DecodeAllFiles/fse-artifact3.bin/fastest-16              1237.51      1259.51      1.02x
BenchmarkDecoder_DecodeAllFiles/fse-artifact3.bin/default-16              1368.29      1393.13      1.02x
BenchmarkDecoder_DecodeAllFiles/fse-artifact3.bin/better-16               1050.82      1039.90      0.99x
BenchmarkDecoder_DecodeAllFiles/fse-artifact3.bin/best-16                 384.69       383.03       1.00x
BenchmarkDecoder_DecodeAllFiles/gettysburg.txt/fastest-16                 325.67       331.85       1.02x
BenchmarkDecoder_DecodeAllFiles/gettysburg.txt/default-16                 210.55       212.73       1.01x
BenchmarkDecoder_DecodeAllFiles/gettysburg.txt/better-16                  210.12       211.09       1.00x
BenchmarkDecoder_DecodeAllFiles/gettysburg.txt/best-16                    213.53       199.22       0.93x
BenchmarkDecoder_DecodeAllFiles/html.txt/fastest-16                       745.09       739.48       0.99x
BenchmarkDecoder_DecodeAllFiles/html.txt/default-16                       724.88       726.15       1.00x
BenchmarkDecoder_DecodeAllFiles/html.txt/better-16                        770.31       769.06       1.00x
BenchmarkDecoder_DecodeAllFiles/html.txt/best-16                          707.72       726.49       1.03x
BenchmarkDecoder_DecodeAllFiles/pi.txt/fastest-16                         10880.58     10859.80     1.00x
BenchmarkDecoder_DecodeAllFiles/pi.txt/default-16                         454.22       455.95       1.00x
BenchmarkDecoder_DecodeAllFiles/pi.txt/better-16                          544.93       547.14       1.00x
BenchmarkDecoder_DecodeAllFiles/pi.txt/best-16                            661.24       654.98       0.99x
BenchmarkDecoder_DecodeAllFiles/pngdata.bin/fastest-16                    1845.59      1874.61      1.02x
BenchmarkDecoder_DecodeAllFiles/pngdata.bin/default-16                    1600.73      1609.12      1.01x
BenchmarkDecoder_DecodeAllFiles/pngdata.bin/better-16                     2006.19      2037.47      1.02x
BenchmarkDecoder_DecodeAllFiles/pngdata.bin/best-16                       1547.17      1554.45      1.00x
BenchmarkDecoder_DecodeAllFiles/sharnd.out/fastest-16                     10884.82     10886.19     1.00x
BenchmarkDecoder_DecodeAllFiles/sharnd.out/default-16                     10892.63     10889.08     1.00x
BenchmarkDecoder_DecodeAllFiles/sharnd.out/better-16                      10898.49     10895.68     1.00x
BenchmarkDecoder_DecodeAllFiles/sharnd.out/best-16                        10867.64     10858.60     1.00x
BenchmarkDecoder_DecodeAllFilesP/Mark.Twain-Tom.Sawyer.txt/fastest-16     3080.40      3121.08      1.01x
BenchmarkDecoder_DecodeAllFilesP/Mark.Twain-Tom.Sawyer.txt/default-16     3007.25      3021.74      1.00x
BenchmarkDecoder_DecodeAllFilesP/Mark.Twain-Tom.Sawyer.txt/better-16      3282.49      3313.83      1.01x
BenchmarkDecoder_DecodeAllFilesP/Mark.Twain-Tom.Sawyer.txt/best-16        3153.04      3131.57      0.99x
BenchmarkDecoder_DecodeAllFilesP/e.txt/fastest-16                         96806.80     96550.57     1.00x
BenchmarkDecoder_DecodeAllFilesP/e.txt/default-16                         2789.45      2815.42      1.01x
BenchmarkDecoder_DecodeAllFilesP/e.txt/better-16                          3559.46      3586.81      1.01x
BenchmarkDecoder_DecodeAllFilesP/e.txt/best-16                            5358.99      5367.39      1.00x
BenchmarkDecoder_DecodeAllFilesP/fse-artifact3.bin/fastest-16             6708.35      5743.08      0.86x
BenchmarkDecoder_DecodeAllFilesP/fse-artifact3.bin/default-16             5677.13      6287.12      1.11x
BenchmarkDecoder_DecodeAllFilesP/fse-artifact3.bin/better-16              5992.62      6444.28      1.08x
BenchmarkDecoder_DecodeAllFilesP/fse-artifact3.bin/best-16                4186.41      4179.85      1.00x
BenchmarkDecoder_DecodeAllFilesP/gettysburg.txt/fastest-16                1673.25      1661.70      0.99x
BenchmarkDecoder_DecodeAllFilesP/gettysburg.txt/default-16                1136.67      1127.95      0.99x
BenchmarkDecoder_DecodeAllFilesP/gettysburg.txt/better-16                 1145.72      1149.76      1.00x
BenchmarkDecoder_DecodeAllFilesP/gettysburg.txt/best-16                   972.79       976.17       1.00x
BenchmarkDecoder_DecodeAllFilesP/html.txt/fastest-16                      3403.28      3399.65      1.00x
BenchmarkDecoder_DecodeAllFilesP/html.txt/default-16                      3297.31      3266.68      0.99x
BenchmarkDecoder_DecodeAllFilesP/html.txt/better-16                       3481.07      3457.98      0.99x
BenchmarkDecoder_DecodeAllFilesP/html.txt/best-16                         3334.76      3300.76      0.99x
BenchmarkDecoder_DecodeAllFilesP/pi.txt/fastest-16                        96573.18     96921.38     1.00x
BenchmarkDecoder_DecodeAllFilesP/pi.txt/default-16                        2811.68      2801.87      1.00x
BenchmarkDecoder_DecodeAllFilesP/pi.txt/better-16                         3649.51      3595.32      0.99x
…